feat: suspense 제거 및 prefetch 적용
