+++ src/apis/queryHooks/chat/useGetLastChat.ts
@@ -10,7 +10,7 @@ function useGetLastChat(roomId: number | null, chatCreate?: string) {
     staleTime: 0,
   });
 
-  const messages = data?.content;
+  const messages = data?.result_data.content;
   const reversedMessages = messages?.map(message => message).reverse();
 
   return { data, error, isLoading, reversedMessages, refetch };
