+++ src/apis/types/basicAuction.ts
@@ -57,7 +57,6 @@ export interface GetBasicAuctionListParams {
   sort?: string;
   page?: number;
   size?: number;
-  isLogin?: boolean;
 }
 
 // ---- 입찰
