+++ src/components/AuctionList/index.tsx
@@ -1,14 +1,11 @@
 'use client';
 
-import { Suspense } from 'react';
-
 import { ArrowRightIcon } from 'lucide-react';
 import Link from 'next/link';
 
 import useGetBasicAuctionList from '@/apis/queryHooks/basicAuction/useGetBasicAuctionList';
 import AuctionCard from '@/components/AuctionCard';
 import ListLayout from '@/components/ListLayout';
-import { useAuth } from '@/provider/authProvider';
 
 import * as S from './AuctionList.css';
 
@@ -20,8 +17,6 @@ export interface AuctionListProps {
 }
 
 export default function AuctionList({ sort, direction, pathname, path }: AuctionListProps) {
-  const { isLoggedIn } = useAuth();
-
   const { data } = useGetBasicAuctionList({
     title: '',
     sort,
@@ -29,7 +24,6 @@ export default function AuctionList({ sort, direction, pathname, path }: Auction
     page: 0,
     size: 8,
     auctionStatus: 'BIDDING',
-    isLogin: isLoggedIn,
   });
 
   if (!data) return null;
@@ -46,21 +40,19 @@ export default function AuctionList({ sort, direction, pathname, path }: Auction
       <div className={S.listWrapper}>
         <ListLayout>
           {data.result_data.content.map(item => (
-            <Suspense key={item.auction_id} fallback={<>AuctionCard</>}>
-              <AuctionCard
-                key={item.auction_id}
-                id={item.auction_id}
-                thumbnailImage={item.thumbnail_path}
-                title={item.title}
-                isLike={item.is_liked}
-                startPrice={item.start_price}
-                startTime={item.start_date}
-                endTime={item.end_date}
-                nowPrice={item.now_price}
-                auctionStatus={item.auction_status}
-                instantBuyPrice={item.instant_buy_price}
-              />
-            </Suspense>
+            <AuctionCard
+              key={item.auction_id}
+              id={item.auction_id}
+              thumbnailImage={item.thumbnail_path}
+              title={item.title}
+              isLike={item.is_liked}
+              startPrice={item.start_price}
+              startTime={item.start_date}
+              endTime={item.end_date}
+              nowPrice={item.now_price}
+              auctionStatus={item.auction_status}
+              instantBuyPrice={item.instant_buy_price}
+            />
           ))}
         </ListLayout>
       </div>
